chore: renderPublic 관련 구조 개선 및 최적화
