+++ src/modules/content/callout/PublicCallout.tsx
@@ -7,17 +7,16 @@ export interface IRenderCalloutProps extends CommonRenderElementProps {
   element: ICallout;
 }
 
-const Callout: React.FC<IRenderCalloutProps> = ({
+const PublicCallout: React.FC<IRenderCalloutProps> = ({
   children,
-  attributes,
   element,
 }) => {
   return (
-    <div className="flex" {...attributes}>
+    <div className="flex">
       <div>icon: {element?.icon}</div>
       <div>{children}</div>
     </div>
   );
 };
 
-export default Callout;
+export default PublicCallout;
