+++ src/modules/content/image/PublicImage.tsx
@@ -1,19 +1,11 @@
 import type { IContentElementProps, IImage } from "@modules/content/types";
 import { useMemo } from "react";
-import { useFocused, useSelected } from "slate-react";
 import { twMerge } from "tailwind-merge";
 
-const Image: React.FC<IContentElementProps<IImage>> = ({
+const PublicImage: React.FC<IContentElementProps<IImage>> = ({
   children,
-  attributes,
   element,
 }) => {
-  const selected = useSelected();
-  const focused = useFocused();
-  const classes = useMemo(() => {
-    return selected && focused ? "ring-indigo-500" : "";
-  }, [focused, selected]);
-
   const style = useMemo(
     () =>
       element.size.type === "FIT"
@@ -26,12 +18,12 @@ const Image: React.FC<IContentElementProps<IImage>> = ({
   );
 
   return (
-    <div {...attributes}>
+    <div>
       <div contentEditable={false} className="relative">
         <img
           src={element.url}
           alt={element.alt}
-          className={twMerge("block max-w-full", classes)}
+          className={twMerge("block max-w-full")}
           style={{ ...style }}
         />
       </div>
@@ -40,4 +32,4 @@ const Image: React.FC<IContentElementProps<IImage>> = ({
   );
 };
 
-export default Image;
+export default PublicImage;
