+++ src/modules/content/list/List.tsx
@@ -1,18 +1,14 @@
-import type { CommonRenderElementProps, IList } from "@modules/content/types";
-import { createElementComponent } from "@modules/editor/custom-slate-editor";
+import type { IContentElementProps, IList } from "@modules/content/types";
 
-export interface IListProps extends CommonRenderElementProps {
-  element: IList;
-}
-
-const { EditorComponent: List, PublicComponent } =
-  createElementComponent<IListProps>(({ children, element, attributes }) => {
-    if (element.ordered) {
-      return <ol {...attributes}>{children}</ol>;
-    }
-    return <ul {...attributes}>{children}</ul>;
-  });
-
-export const PublicList = PublicComponent;
+const List: React.FC<IContentElementProps<IList>> = ({
+  children,
+  element,
+  attributes,
+}) => {
+  if (element.ordered) {
+    return <ol {...attributes}>{children}</ol>;
+  }
+  return <ul {...attributes}>{children}</ul>;
+};
 
 export default List;
