+++ src/modules/content/paragraph/Paragraph.tsx
@@ -1,13 +1,16 @@
-import type { CommonRenderElementProps } from "@modules/content/types";
-import { createElementComponent } from "@modules/editor/custom-slate-editor";
+import type {
+  CommonRenderElementProps,
+  IContentElementProps,
+  IParagraph,
+} from "@modules/content/types";
 
 export type IParagraphProps = CommonRenderElementProps;
 
-const { EditorComponent: Paragraph, PublicComponent } =
-  createElementComponent<IParagraphProps>(({ children, attributes }) => {
-    return <p {...attributes}>{children}</p>;
-  });
-
-export const PublicParagraph = PublicComponent;
+const Paragraph: React.FC<IContentElementProps<IParagraph>> = ({
+  attributes,
+  children,
+}) => {
+  return <p {...attributes}>{children}</p>;
+};
 
 export default Paragraph;
