+++ src/modules/layout/AdminLayout.tsx
@@ -1,7 +1,30 @@
+import type { LayoutWrapper } from "@modules/content/types";
+import { anonClient } from "@modules/supabase/supabase";
 import Link from "next/link";
-import type { ReactNode } from "react";
+import { useEffect } from "react";
 
-export default function adminLayoutWrapper(page: ReactNode) {
+const AdminLayoutWrapper: LayoutWrapper = ({ page }) => {
+  useEffect(() => {
+    const { data, error } = anonClient.auth.onAuthStateChange(
+      (event, session) => {
+        fetch("/api/auth", {
+          method: "POST",
+          headers: new Headers({ "Content-Type": "application/json" }),
+          credentials: "same-origin",
+          body: JSON.stringify({ event, session }),
+        });
+      }
+    );
+    if (error) {
+      // eslint-disable-next-line no-console
+      console.error("supabase.auth.onAuthStateChange", error);
+    }
+
+    return () => {
+      data?.unsubscribe();
+    };
+  }, []);
+
   return (
     <div className="flex mx-auto">
       <div className="w-64 p-8">
@@ -14,4 +37,6 @@ export default function adminLayoutWrapper(page: ReactNode) {
       <main className="flex-1 m-8">{page}</main>
     </div>
   );
-}
+};
+
+export default AdminLayoutWrapper;
