+++ src/pages/_app.tsx
@@ -3,11 +3,10 @@ import { useMyStoreMemo } from "@common/store";
 import { config as fontAwesomeConfig } from "@fortawesome/fontawesome-svg-core";
 import type { MonnomlogPage } from "@modules/content/types";
 import Default from "@modules/layout/Default";
-import { anonClient } from "@modules/supabase/supabase";
 import { DefaultSeo } from "next-seo";
 import type { AppProps } from "next/app";
 import Head from "next/head";
-import type { ReactElement } from "react";
+import type { ReactNode } from "react";
 import { useEffect } from "react";
 
 import defaultSEOConfig from "../../next-seo.config";
@@ -17,35 +16,16 @@ import "@common/globals.css";
 
 fontAwesomeConfig.autoAddCss = false;
 
-const defaultGetLayout = (page: ReactElement) => <Default>{page}</Default>;
+const defaultGetLayout = (props: { page: ReactNode }) => (
+  <Default>{props.page}</Default>
+);
 
 interface IMyAppProps extends AppProps {
   Component: MonnomlogPage;
 }
 
 const MyApp = ({ Component, pageProps }: IMyAppProps) => {
-  useEffect(() => {
-    const { data, error } = anonClient.auth.onAuthStateChange(
-      (event, session) => {
-        fetch("/api/auth", {
-          method: "POST",
-          headers: new Headers({ "Content-Type": "application/json" }),
-          credentials: "same-origin",
-          body: JSON.stringify({ event, session }),
-        });
-      }
-    );
-    if (error) {
-      // eslint-disable-next-line no-console
-      console.error("supabase.auth.onAuthStateChange", error);
-    }
-
-    return () => {
-      data?.unsubscribe();
-    };
-  }, []);
-
-  const wrap = Component.layoutWrapper ?? defaultGetLayout;
+  const LayoutWrapper = Component.layoutWrapper ?? defaultGetLayout;
 
   const colorMode = useMyStoreMemo((store) => {
     return store.colorMode;
@@ -91,7 +71,7 @@ const MyApp = ({ Component, pageProps }: IMyAppProps) => {
         />
       </Head>
       <DefaultSeo {...defaultSEOConfig} />
-      {wrap(<Component {...pageProps} />)}
+      <LayoutWrapper page={<Component {...pageProps} />} />
     </>
   );
 };
